Removed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@9710 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/physics/physical_object.hpp
@@ -70,7 +70,7 @@ public:
                  PhysicalObject (const XMLNode &node);
     
                  PhysicalObject(const std::string& model,
-                                bodyTypes shape, int mass, int radius,
+                                bodyTypes shape, float mass, float radius,
                                 const core::vector3df& hpr,
                                 const core::vector3df& pos,
                                 const core::vector3df& scale);
